refactor(units): SI temperature header (#4)
