+++ posu/units/si/si.hpp
@@ -4,96 +4,10 @@
 #include "posu/units/si/electric_current.hpp"
 #include "posu/units/si/length.hpp"
 #include "posu/units/si/mass.hpp"
+#include "posu/units/si/thermodynamic_temperature.hpp"
 #include "posu/units/si/time.hpp"
 
 namespace posu::units {
-
-    struct thermodynamic_temperature_tag {
-        constexpr thermodynamic_temperature_tag() noexcept = default;
-    };
-
-    template<typename Rep, typename Period>
-    using thermodynamic_temperature = base_unit<Rep, Period, thermodynamic_temperature_tag>;
-
-    using attokelvins  = thermodynamic_temperature<int, std::atto>;
-    using femtokelvins = thermodynamic_temperature<int, std::femto>;
-    using picokelvins  = thermodynamic_temperature<int, std::pico>;
-    using nanokelvins  = thermodynamic_temperature<int, std::nano>;
-    using microkelvins = thermodynamic_temperature<int, std::micro>;
-    using millikelvins = thermodynamic_temperature<int, std::milli>;
-    using centikelvins = thermodynamic_temperature<int, std::centi>;
-    using decikelvins  = thermodynamic_temperature<int, std::deci>;
-    using kelvins      = thermodynamic_temperature<int, std::ratio<1>>;
-    using decakelvins  = thermodynamic_temperature<int, std::deca>;
-    using hectokelvins = thermodynamic_temperature<int, std::hecto>;
-    using kilokelvins  = thermodynamic_temperature<int, std::kilo>;
-    using megakelvins  = thermodynamic_temperature<int, std::mega>;
-    using gigakelvins  = thermodynamic_temperature<int, std::giga>;
-    using terakelvins  = thermodynamic_temperature<int, std::tera>;
-    using petakelvins  = thermodynamic_temperature<int, std::peta>;
-    using exakelvins   = thermodynamic_temperature<int, std::exa>;
-
-    inline namespace literals {
-
-        inline namespace thermodynamic_temperature_literals {
-
-            [[nodiscard]] constexpr auto operator""_aK(unsigned long long value) -> attokelvins;
-            [[nodiscard]] constexpr auto operator""_aK(long double value)
-                -> thermodynamic_temperature<double, std::atto>;
-            [[nodiscard]] constexpr auto operator""_fK(unsigned long long value) -> femtokelvins;
-            [[nodiscard]] constexpr auto operator""_fK(long double value)
-                -> thermodynamic_temperature<double, std::femto>;
-            [[nodiscard]] constexpr auto operator""_pK(unsigned long long value) -> picokelvins;
-            [[nodiscard]] constexpr auto operator""_pK(long double value)
-                -> thermodynamic_temperature<double, std::pico>;
-            [[nodiscard]] constexpr auto operator""_nK(unsigned long long value) -> nanokelvins;
-            [[nodiscard]] constexpr auto operator""_nK(long double value)
-                -> thermodynamic_temperature<double, std::nano>;
-            [[nodiscard]] constexpr auto operator""_uK(unsigned long long value) -> microkelvins;
-            [[nodiscard]] constexpr auto operator""_uK(long double value)
-                -> thermodynamic_temperature<double, std::micro>;
-            [[nodiscard]] constexpr auto operator""_mK(unsigned long long value) -> millikelvins;
-            [[nodiscard]] constexpr auto operator""_mK(long double value)
-                -> thermodynamic_temperature<double, std::milli>;
-            [[nodiscard]] constexpr auto operator""_cK(unsigned long long value) -> centikelvins;
-            [[nodiscard]] constexpr auto operator""_cK(long double value)
-                -> thermodynamic_temperature<double, std::centi>;
-            [[nodiscard]] constexpr auto operator""_dK(unsigned long long value) -> decikelvins;
-            [[nodiscard]] constexpr auto operator""_dK(long double value)
-                -> thermodynamic_temperature<double, std::deci>;
-            [[nodiscard]] constexpr auto operator""_K(unsigned long long value) -> kelvins;
-            [[nodiscard]] constexpr auto operator""_K(long double value)
-                -> thermodynamic_temperature<double, std::ratio<1>>;
-            [[nodiscard]] constexpr auto operator""_daK(unsigned long long value) -> decakelvins;
-            [[nodiscard]] constexpr auto operator""_daK(long double value)
-                -> thermodynamic_temperature<double, std::deca>;
-            [[nodiscard]] constexpr auto operator""_hK(unsigned long long value) -> hectokelvins;
-            [[nodiscard]] constexpr auto operator""_hK(long double value)
-                -> thermodynamic_temperature<double, std::hecto>;
-            [[nodiscard]] constexpr auto operator""_kK(unsigned long long value) -> kilokelvins;
-            [[nodiscard]] constexpr auto operator""_kK(long double value)
-                -> thermodynamic_temperature<double, std::kilo>;
-            [[nodiscard]] constexpr auto operator""_MK(unsigned long long value) -> megakelvins;
-            [[nodiscard]] constexpr auto operator""_MK(long double value)
-                -> thermodynamic_temperature<double, std::mega>;
-            [[nodiscard]] constexpr auto operator""_GK(unsigned long long value) -> gigakelvins;
-            [[nodiscard]] constexpr auto operator""_GK(long double value)
-                -> thermodynamic_temperature<double, std::giga>;
-            [[nodiscard]] constexpr auto operator""_TK(unsigned long long value) -> terakelvins;
-            [[nodiscard]] constexpr auto operator""_TK(long double value)
-                -> thermodynamic_temperature<double, std::tera>;
-            [[nodiscard]] constexpr auto operator""_PK(unsigned long long value) -> petakelvins;
-            [[nodiscard]] constexpr auto operator""_PK(long double value)
-                -> thermodynamic_temperature<double, std::peta>;
-            [[nodiscard]] constexpr auto operator""_EK(unsigned long long value) -> exakelvins;
-            [[nodiscard]] constexpr auto operator""_EK(long double value)
-                -> thermodynamic_temperature<double, std::exa>;
-
-        } // namespace thermodynamic_temperature_literals
-
-    } // namespace literals
-
-    using namespace literals::thermodynamic_temperature_literals;
 
     struct amount_of_substance_tag {
         constexpr amount_of_substance_tag() noexcept = default;
